Еще структурировал проект Переместил модели DTO в отдельную папку и поправил namespaces
