+++ MetricsAgent/Controllers/Models/CpuMetricDto.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace MetricsAgent.Controllers.Models
+{
+    public class CpuMetricDto
+    {
+        public int Id { get; set; }
+        public int Value { get; set; }
+        public DateTimeOffset Time { get; set; }
+    }
+}
