+++ MetricsAgent/Controllers/Models/HddMetricDto.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace MetricsAgent.Controllers.Responses
+{
+    public class HddMetricDto
+    {
+        public int Id { get; set; }
+        public int Value { get; set; }
+        public DateTimeOffset Time { get; set; }
+    }
+}
