+++ MetricsAgent/Controllers/Responses/CpuMetricsResponse.cs
@@ -1,4 +1,4 @@
-﻿using System;
+﻿using MetricsAgent.Controllers.Models;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -8,12 +8,5 @@ namespace MetricsAgent.Controllers.Responses
     public class CpuMetricsResponse
     {
         public List<CpuMetricDto> Metrics { get; set; }
-    }
-
-    public class CpuMetricDto
-    {
-        public int Id { get; set; }
-        public int Value { get; set; }
-        public DateTimeOffset Time { get; set; }
     }
 }
