+++ MetricsAgent/Controllers/Responses/DotNetMetricsResponse.cs
@@ -1,4 +1,4 @@
-﻿using System;
+﻿using MetricsAgent.Controllers.Models;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -8,11 +8,5 @@ namespace MetricsAgent.Controllers.Responses
     public class DotNetMetricsResponse
     {
         public List<DotNetMetricDto> Metrics { get; set; }
-    }
-    public class DotNetMetricDto
-    {
-        public int Id { get; set; }
-        public int Value { get; set; }
-        public DateTimeOffset Time { get; set; }
     }
 }
