+++ MetricsAgent/Controllers/Responses/NetworkMetricsResponse.cs
@@ -1,19 +1,12 @@
-﻿using System;
+﻿using MetricsAgent.Controllers.Responses;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MetricsAgent.Controllers.Responses
+namespace MetricsAgent.Controllers.Models
 {
     public class NetworkMetricsResponse
     {
         public List<NetworkMetricDto> Metrics { get; set; }
-    }
-
-    public class NetworkMetricDto
-    {
-        public int Id { get; set; }
-        public int Value { get; set; }
-        public DateTimeOffset Time { get; set; }
     }
 }
