+++ MetricsAgent/Controllers/Responses/RamMetricsResponse.cs
@@ -1,19 +1,12 @@
-﻿using System;
+﻿using MetricsAgent.Controllers.Responses;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MetricsAgent.Controllers.Responses
+namespace MetricsAgent.Controllers.Models
 {
     public class RamMetricsResponse
     {
         public List<RamMetricDto> Metrics { get; set; }
-    }
-
-    public class RamMetricDto
-    {
-        public int Id { get; set; }
-        public int Value { get; set; }
-        public DateTimeOffset Time { get; set; }
     }
 }
