+++ MetricsAgentTests/DotNetMetricsControllerTests.cs
@@ -1,6 +1,6 @@
 ﻿using MetricsAgent.Controllers;
+using MetricsAgent.DAL.Interfaces;
 using MetricsAgent.DAL.Models;
-using MetricsAgent.DAL.Repositories;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using Moq;
